+++ Sources/Internal/Render/Highlevel/Vegetation/VegetationFixedGeometry.cpp
@@ -269,7 +269,7 @@ void VegetationFixedGeometry::GenerateVertices(uint32 maxClusters,
     uint32 totalTiles = tilesPerRow * tilesPerRow;
     for(uint32 layerIndex = 0; layerIndex < maxLayerTypes; ++layerIndex)
     {
-        layerOffsets[layerIndex] = vertexIndex;
+        layerOffsets[layerIndex] = static_cast<uint32>(vertexIndex);
         
         TextureSheetCell& cellData = textureSheet.cells[layerIndex];
         
@@ -308,7 +308,7 @@ void VegetationFixedGeometry::GenerateVertices(uint32 maxClusters,
         for(size_t clusterIndex = 0; clusterIndex < maxTotalClusters; ++clusterIndex)
         {
             uint32 clusterIndexX = clusterIndex % maxClusterRowSize;
-            uint32 clusterIndexY = clusterIndex / maxClusterRowSize;
+            uint32 clusterIndexY = static_cast<uint32>(clusterIndex / maxClusterRowSize);
             
             uint32 matrixIndex = (clusterIndexX / maxClusters) + tilesPerRow * (clusterIndexY / maxClusters); //0...15
             DVASSERT(matrixIndex < (tilesPerRow * tilesPerRow));
@@ -449,7 +449,7 @@ void VegetationFixedGeometry::PrepareIndexBufferData(uint32 indexBufferIndex,
                     
                     indexBufferBBox.AddPoint(vertex.coord);
                     
-                    preparedIndices.push_back(vertexIndex);
+                    preparedIndices.push_back(static_cast<int32>(vertexIndex));
                 }
             }
         }
@@ -523,7 +523,7 @@ void VegetationFixedGeometry::PrepareSortedIndexBufferVariations(size_t& current
         }
         
         RenderDataObject* indexBuffer = new RenderDataObject();
-        indexBuffer->SetIndices(VEGETATION_INDEX_TYPE, (uint8*)(&indexData[prevIndexIndex]), (currentIndexIndex - prevIndexIndex));
+        indexBuffer->SetIndices(VEGETATION_INDEX_TYPE, (uint8*)(&indexData[prevIndexIndex]), static_cast<int32>(currentIndexIndex - prevIndexIndex));
         
         VegetationSortedBufferItem sortedBufferItem;
         sortedBufferItem.SetRenderDataObject(indexBuffer);
@@ -549,7 +549,7 @@ void VegetationFixedGeometry::GenerateRenderDataObjects(VegetationRenderData& re
     vertexRenderDataObject->SetStream(EVF_BINORMAL, TYPE_FLOAT, 3, sizeof(VegetationVertex), &(vertexData[0].binormal));
     vertexRenderDataObject->SetStream(EVF_TANGENT, TYPE_FLOAT, 3, sizeof(VegetationVertex), &(vertexData[0].tangent));
     vertexRenderDataObject->SetStream(EVF_TEXCOORD0, TYPE_FLOAT, 2, sizeof(VegetationVertex), &(vertexData[0].texCoord0));
-    vertexRenderDataObject->BuildVertexBuffer(vertexData.size(), BDT_STATIC_DRAW, true);
+    vertexRenderDataObject->BuildVertexBuffer(static_cast<uint32>(vertexData.size()), BDT_STATIC_DRAW, true);
     
     size_t totalIndexObjectArrayCount = indexRenderDataObject.size();
     for(size_t indexArrayIndex = 0; indexArrayIndex < totalIndexObjectArrayCount; ++indexArrayIndex)
