+++ Sources/Internal/Scene3D/Systems/Controller/SnapToLandscapeControllerSystem.cpp
@@ -63,7 +63,7 @@ void SnapToLandscapeControllerSystem::AddEntity(Entity * entity)
 
 void SnapToLandscapeControllerSystem::RemoveEntity(Entity * entity)
 {
-    uint32 size = entities.size();
+    uint32 size = static_cast<uint32>(entities.size());
     for(uint32 i = 0; i < size; ++i)
     {
         if(entities[i] == entity)
@@ -88,7 +88,7 @@ void SnapToLandscapeControllerSystem::RemoveEntity(Entity * entity)
 
 void SnapToLandscapeControllerSystem::Process(float32 timeElapsed)
 {
-    const uint32 size = entities.size();
+    const uint32 size = static_cast<uint32>(entities.size());
     if(0 == size) return;
     
     Landscape *landscape = FindLandscape(GetScene());
